:truck: [#4608] Move Objects API admin fields to generic location
previously these were defined as part of registrations, but since they are now also used by prefill, their location should reflect that they are generic

* ObjectsAPIGroup
* ObjectTypeSelect
* ObjectTypeVersionSelect
