+++ src/openforms/js/components/admin/form_design/registrations/objectsapi/fields/index.js
@@ -1,6 +1,3 @@
-export {default as ObjectsAPIGroup} from './ObjectsAPIGroup';
-export {default as ObjectTypeSelect} from './ObjectTypeSelect';
-export {default as ObjectTypeVersionSelect} from './ObjectTypeVersionSelect';
 export {LegacyDocumentTypesFieldet} from './LegacyDocumentType';
 export {DocumentTypesFieldet} from './DocumentTypes';
 export {default as UpdateExistingObject} from './UpdateExistingObject';
