+++ src/openforms/js/components/admin/forms/objects_api/index.js
@@ -1,0 +1,3 @@
+export {default as ObjectsAPIGroup} from './ObjectsAPIGroup';
+export {default as ObjectTypeSelect} from './ObjectTypeSelect';
+export {default as ObjectTypeVersionSelect} from './ObjectTypeVersionSelect';
